GlobalSensitivity.jl: Robust, Fast, and Parallel Global Sensitivity Analysis (GSA) in Julia

Global Sensitivity Analysis (GSA) methods are used to quantify the uncertainty in the output of a model with respect to the parameters. These methods allow practitioners to measure both parameter's individual contributions and the contribution of their interactions to the output uncertainty.

Installation

To use this functionality, you must install GlobalSensitivity.jl:

]add GlobalSensitivity
using GlobalSensitivity

Note: GlobalSensitivity.jl is unrelated to the GlobalSensitivityAnalysis.jl package.

General Interface

The general interface for performing global sensitivity analysis using this package is:

GlobalSensitivity.gsa — Method

gsa(f, method::GSAMethod, param_range; samples, batch=false)

where:

y=f(x) is a function that takes in a single vector and spits out a single vector or scalar. If batch=true, then f takes in a matrix where each row is a set of parameters, and returns a matrix where each row is the output for the corresponding row of parameters.

method is one of the available GSA methods.

param_range is a vector of tuples for the upper and lower bound for the given parameter i.

samples is a required keyword argument for the number of samples of parameters for the design matrix. Note that this is not relevant for Fractional Factorial Method and Morris Method.

Additionally,

For Delta Moment-Independent Method, EASI Method and Regression Method input and output matrix-based method as follows is available:

res = gsa(X, Y, method)

where:

X is the number of parameters * samples matrix with parameter values.

Y is the output dimension * number of samples matrix with are evaluated at X's columns.

method is one of the GSA methods below.

For Sobol Method, one can use the following design matrices-based method instead of parameter range-based method discussed earlier:

effects = gsa(f, method, A, B; batch=false)

where A and B are design matrices, with each row being a set of parameters. Note that generate_design_matrices from QuasiMonteCarlo.jl can be used to generate the design matrices.

The descriptions of the available methods can be found in the Methods section. The gsa interface allows for utilizing batched functions with the batch kwarg discussed above for parallel computation of GSA results.

Citing

If you use this software in your work, please cite:

@article{dixit2022globalsensitivity,
  title={GlobalSensitivity. jl: Performant and Parallel Global Sensitivity Analysis with Julia},
  author={Dixit, Vaibhav Kumar and Rackauckas, Christopher},
  journal={Journal of Open Source Software},
  volume={7},
  number={76},
  pages={4561},
  year={2022}
}

Reproducibility

The documentation of this SciML package was built using these direct dependencies,

Status `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl/docs/Project.toml`
  [13f3f980] CairoMakie v0.10.6
  [ae264745] Copulas v0.1.10
  [31c24e10] Distributions v0.25.98
  [e30172f5] Documenter v0.27.25
  [587475ba] Flux v0.13.17
  [af5da776] GlobalSensitivity v2.2.1 `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl`
  [7f7a1694] Optimization v3.15.2
  [42dfb2eb] OptimizationOptimisers v0.1.4
  [1dea7af3] OrdinaryDiffEq v6.53.3
  [91a5bcdd] Plots v1.38.16
  [8a4e6c94] QuasiMonteCarlo v0.3.1
  [1ed8b502] SciMLSensitivity v7.34.0
  [860ef19b] StableRNGs v1.0.0
  [10745b16] Statistics v1.9.0

and using this machine and Julia version.

Julia Version 1.9.2
Commit e4ee485e909 (2023-07-05 09:39 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 2 × Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, skylake-avx512)
  Threads: 1 on 2 virtual cores

A more complete overview of all dependencies and their versions is also provided.

Status `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl/docs/Manifest.toml`
  [47edcb42] ADTypes v0.1.5
  [a4c015fc] ANSIColoredPrinters v0.0.1
  [621f4979] AbstractFFTs v1.4.0
  [1520ce14] AbstractTrees v0.4.4
  [7d9f7c33] Accessors v0.1.32
  [79e6a3ab] Adapt v3.6.2
  [f20549b4] AlphaStableDistributions v1.1.6
  [27a7e980] Animations v0.4.1
  [dce04be8] ArgCheck v2.3.0
  [ec485272] ArnoldiMethod v0.2.0
  [4fba245c] ArrayInterface v7.4.11
  [30b0a656] ArrayInterfaceCore v0.1.29
  [a9b6321e] Atomix v0.1.0
  [67c07d97] Automa v0.8.3
  [13072b0f] AxisAlgorithms v1.0.1
  [39de3d68] AxisArrays v0.4.6
  [ab4f0b2a] BFloat16s v0.4.2
  [198e06fe] BangBang v0.3.39
  [9718e550] Baselet v0.1.1
  [d1d4a3ce] BitFlags v0.1.7
  [62783981] BitTwiddlingConvenienceFunctions v0.1.5
  [fa961155] CEnum v0.4.2
  [2a0fbf3d] CPUSummary v0.2.3
  [96374032] CRlibm v1.0.1
⌅ [052768ef] CUDA v4.2.0
  [1af6417a] CUDA_Runtime_Discovery v0.2.2
  [159f3aea] Cairo v1.0.5
  [13f3f980] CairoMakie v0.10.6
  [49dc2e85] Calculus v0.5.1
  [7057c7e9] Cassette v0.3.11
  [082447d4] ChainRules v1.52.0
  [d360d2e6] ChainRulesCore v1.16.0
  [fb6a15b2] CloseOpenIntervals v0.1.12
  [944b1d66] CodecZlib v0.7.1
  [a2cac450] ColorBrewer v0.4.0
  [35d6a980] ColorSchemes v3.21.0
  [3da002f7] ColorTypes v0.11.4
  [c3611d14] ColorVectorSpace v0.9.10
  [5ae59095] Colors v0.12.10
  [861a8166] Combinatorics v1.0.2
  [38540f10] CommonSolve v0.2.4
  [bbf7d656] CommonSubexpressions v0.3.0
  [34da2185] Compat v4.7.0
  [a33af91c] CompositionsBase v0.1.2
  [2569d6c7] ConcreteStructs v0.2.3
  [f0e56b4a] ConcurrentUtilities v2.2.0
  [88cd18e8] ConsoleProgressMonitor v0.1.2
  [187b0558] ConstructionBase v1.5.2
  [6add18c4] ContextVariablesX v0.1.3
  [d38c429a] Contour v0.6.2
  [ae264745] Copulas v0.1.10
  [adafc99b] CpuId v0.3.1
  [667455a9] Cubature v1.5.1
  [717857b8] DSP v0.7.8
  [9a962f9c] DataAPI v1.15.0
  [864edb3b] DataStructures v0.18.14
  [e2d170a0] DataValueInterfaces v1.0.0
  [244e2a9f] DefineSingletons v0.1.2
  [8bb1440f] DelimitedFiles v1.9.1
  [2b5f629d] DiffEqBase v6.126.0
  [459566f4] DiffEqCallbacks v2.26.1
  [77a26b50] DiffEqNoiseProcess v5.17.2
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.15.1
  [b4f34e82] Distances v0.10.8
  [31c24e10] Distributions v0.25.98
  [ffbed154] DocStringExtensions v0.9.3
  [e30172f5] Documenter v0.27.25
  [fa6b7ba4] DualNumbers v0.6.8
  [da5c29d0] EllipsisNotation v1.7.0
  [4e289a0a] EnumX v1.0.4
⌅ [7da242da] Enzyme v0.11.0
⌅ [f151be2c] EnzymeCore v0.3.0
  [90fa49ef] ErrorfreeArithmetic v0.5.2
  [460bff9d] ExceptionUnwrapping v0.1.9
  [d4d017d3] ExponentialUtilities v1.24.0
  [e2ba6199] ExprTools v0.1.9
  [411431e0] Extents v0.1.1
  [c87230d0] FFMPEG v0.4.1
  [7a1cc6ca] FFTW v1.7.1
  [cc61a311] FLoops v0.2.1
  [b9860ae5] FLoopsBase v0.1.1
  [7034ab61] FastBroadcast v0.2.5
  [9aa1b823] FastClosures v0.3.2
  [29a986be] FastLapackInterface v2.0.0
  [fa42c844] FastRounding v0.3.1
  [5789e2e9] FileIO v1.16.1
  [1a297f60] FillArrays v1.3.0
  [6a86dc24] FiniteDiff v2.21.1
  [53c48c17] FixedPointNumbers v0.8.4
  [587475ba] Flux v0.13.17
  [59287772] Formatting v0.4.2
  [f6369f11] ForwardDiff v0.10.35
  [b38be410] FreeType v4.0.0
  [663a7486] FreeTypeAbstraction v0.10.0
  [069b7b12] FunctionWrappers v1.1.3
  [77dc65aa] FunctionWrappersWrappers v0.1.3
  [d9f16b24] Functors v0.4.4
  [0c68f7d7] GPUArrays v8.8.1
  [46192b85] GPUArraysCore v0.1.5
⌅ [61eb1bfa] GPUCompiler v0.19.4
  [28b8d3ca] GR v0.72.7
  [92c85e6c] GSL v1.0.1
  [c145ed77] GenericSchur v0.5.3
  [cf35fbd7] GeoInterface v1.3.1
  [5c1252a2] GeometryBasics v0.4.7
  [af5da776] GlobalSensitivity v2.2.1 `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl`
  [a2bd30eb] Graphics v1.1.2
  [86223c79] Graphs v1.8.0
  [3955a311] GridLayoutBase v0.9.1
  [42e2da0e] Grisu v1.0.2
  [cd3eb016] HTTP v1.9.8
  [3e5b6fbb] HostCPUFeatures v0.1.15
  [34004b35] HypergeometricFunctions v0.3.17
  [b5f81e59] IOCapture v0.2.3
  [7869d1d1] IRTools v0.4.10
  [615f187c] IfElse v0.1.1
  [2803e5a7] ImageAxes v0.6.10
  [c817782e] ImageBase v0.1.5
  [a09fc81d] ImageCore v0.9.4
  [82e4d734] ImageIO v0.6.6
  [bc367c6b] ImageMetadata v0.9.8
  [9b13fd28] IndirectArrays v1.0.0
  [d25df0c9] Inflate v0.1.3
  [22cec73e] InitialValues v0.3.1
  [18e54dd8] IntegerMathUtils v0.1.2
  [a98d9a8b] Interpolations v0.14.7
  [d1acc4aa] IntervalArithmetic v0.20.9
  [8197267c] IntervalSets v0.7.4
  [3587e190] InverseFunctions v0.1.10
  [92d709cd] IrrationalConstants v0.2.2
  [f1662d9f] Isoband v0.1.1
  [c8e1da08] IterTools v1.8.0
  [82899510] IteratorInterfaceExtensions v1.0.0
  [033835bb] JLD2 v0.4.31
  [1019f520] JLFzf v0.1.5
  [692b3bcd] JLLWrappers v1.4.1
  [682c06a0] JSON v0.21.4
  [b835a17e] JpegTurbo v0.1.2
  [b14d175d] JuliaVariables v0.2.4
  [ccbc3e58] JumpProcesses v9.6.3
  [ef3ab10e] KLU v0.4.0
  [63c18a36] KernelAbstractions v0.9.6
  [5ab0869b] KernelDensity v0.6.7
  [ba0b0d4f] Krylov v0.9.1
⌅ [929cbde3] LLVM v5.2.0
  [b964fa9f] LaTeXStrings v1.3.0
  [23fbe1c1] Latexify v0.16.1
  [73f95e8e] LatticeRules v0.0.1
  [10f19ff3] LayoutPointers v0.1.14
  [50d2b5c4] Lazy v0.15.1
  [8cdb02fc] LazyModules v0.3.1
  [1d6d02ad] LeftChildRightSiblingTrees v0.2.0
  [2d8b4e74] LevyArea v1.0.0
  [d3d80556] LineSearches v7.2.0
  [7ed4a6bd] LinearSolve v2.3.0
  [2ab3a3ac] LogExpFunctions v0.3.24
  [e6f89c97] LoggingExtras v1.0.0
  [bdcacae8] LoopVectorization v0.12.162
  [d8e11817] MLStyle v0.4.17
  [f1d291b0] MLUtils v0.4.3
  [1914dd2f] MacroTools v0.5.10
  [ee78f7c6] Makie v0.19.6
  [20f20a25] MakieCore v0.6.3
  [d125e4d3] ManualMemory v0.1.8
  [dbb5928d] MappedArrays v0.4.2
  [7eb4fadd] Match v1.2.0
  [0a4f8689] MathTeXEngine v0.5.6
  [739be429] MbedTLS v1.1.7
  [442fdcdd] Measures v0.3.2
  [128add7d] MicroCollections v0.1.4
  [978d7f02] MiniQhull v0.4.0
  [e1d29d7a] Missings v1.1.0
  [e94cdb99] MosaicViews v0.3.4
  [46d2c3a1] MuladdMacro v0.2.4
  [37188c8d] MvNormalCDF v0.2.6
  [d41bc354] NLSolversBase v7.8.3
  [2774e3e8] NLsolve v4.5.1
⌅ [872c559c] NNlib v0.8.21
  [a00861dc] NNlibCUDA v0.2.7
  [77ba4419] NaNMath v1.0.2
  [71a1bf82] NameResolution v0.1.5
  [f09324ee] Netpbm v1.1.0
  [8913a72c] NonlinearSolve v1.8.0
⌅ [d8793406] ObjectFile v0.3.7
  [510215fc] Observables v0.5.4
  [6fe1bfb0] OffsetArrays v1.12.10
  [0b1bfda6] OneHotArrays v0.2.4
  [52e1d378] OpenEXR v0.3.2
  [4d8831e6] OpenSSL v1.4.1
  [429524aa] Optim v1.7.6
  [3bd65402] Optimisers v0.2.18
  [7f7a1694] Optimization v3.15.2
  [42dfb2eb] OptimizationOptimisers v0.1.4
  [bac558e1] OrderedCollections v1.6.0
  [1dea7af3] OrdinaryDiffEq v6.53.3
  [90014a1f] PDMats v0.11.17
  [f57f5aa1] PNGFiles v0.3.17
  [19eb6ba3] Packing v0.5.0
  [5432bcbf] PaddedViews v0.5.12
  [d96e819e] Parameters v0.12.3
  [69de0a69] Parsers v2.7.1
  [b98c9c47] Pipe v1.3.0
  [eebad327] PkgVersion v0.3.2
  [ccf2f8ad] PlotThemes v3.1.0
  [995b91a9] PlotUtils v1.3.5
  [91a5bcdd] Plots v1.38.16
  [e409e4f3] PoissonRandom v0.4.4
  [f517fe37] Polyester v0.7.3
  [1d0040c9] PolyesterWeave v0.2.1
  [647866c9] PolygonOps v0.1.2
  [f27b6e38] Polynomials v3.2.13
  [85a6dd25] PositiveFactorizations v0.2.4
  [d236fae5] PreallocationTools v0.4.12
  [aea7be01] PrecompileTools v1.1.2
  [21216c6a] Preferences v1.4.0
  [8162dcfd] PrettyPrint v0.2.0
  [27ebfcd6] Primes v0.5.3
  [33c8b6b6] ProgressLogging v0.1.4
  [92933f4c] ProgressMeter v1.7.2
  [4b34888f] QOI v1.0.0
  [1fd47b50] QuadGK v2.8.2
  [8a4e6c94] QuasiMonteCarlo v0.3.1
  [74087812] Random123 v1.6.1
  [e6cf234a] RandomNumbers v1.5.3
  [b3c3ace0] RangeArrays v0.3.2
  [c84ed2f1] Ratios v0.4.5
  [c1ae055f] RealDot v0.1.0
  [3cdcf5f2] RecipesBase v1.3.4
  [01d81517] RecipesPipeline v0.6.12
  [731186ca] RecursiveArrayTools v2.38.5
  [f2c3362d] RecursiveFactorization v0.2.18
  [189a3867] Reexport v1.2.2
  [42d2dcc6] Referenceables v0.1.2
  [05181044] RelocatableFolders v1.0.0
  [ae029012] Requires v1.3.0
  [ae5879a3] ResettableStacks v1.1.1
  [37e2e3b7] ReverseDiff v1.14.6
  [79098fc4] Rmath v0.7.1
  [f2b01f46] Roots v2.0.17
  [5eaf0fd0] RoundingEmulator v0.2.1
  [7e49a35a] RuntimeGeneratedFunctions v0.5.11
  [fdea26ae] SIMD v3.4.5
  [94e857df] SIMDTypes v0.1.0
  [476501e8] SLEEFPirates v0.6.39
  [7b38b023] ScanByte v0.4.0
  [0bca4576] SciMLBase v1.93.0
  [e9a6253c] SciMLNLSolve v0.1.8
  [c0aeaf25] SciMLOperators v0.3.3
  [1ed8b502] SciMLSensitivity v7.34.0
  [6c6a2e73] Scratch v1.2.0
  [3cc68bcd] SetRounding v0.2.1
  [efcf1570] Setfield v1.1.1
  [605ecd9f] ShowCases v0.1.0
  [992d4aef] Showoff v1.0.3
  [73760f76] SignedDistanceFields v0.4.0
  [777ac1f9] SimpleBufferStream v1.1.0
  [727e6d20] SimpleNonlinearSolve v0.1.16
  [699a6c99] SimpleTraits v0.9.4
  [ce78b400] SimpleUnPack v1.1.0
  [45858cf5] Sixel v0.1.2
  [66db9d55] SnoopPrecompile v1.0.3
  [ed01d8cd] Sobol v1.5.0
  [a2af1166] SortingAlgorithms v1.1.1
  [47a9eef4] SparseDiffTools v2.4.1
  [e56a9233] Sparspak v0.3.9
  [276daf66] SpecialFunctions v2.3.0
  [171d559e] SplittablesBase v0.1.15
  [c5dd0088] StableHashTraits v0.3.1
  [860ef19b] StableRNGs v1.0.0
  [cae243ae] StackViews v0.1.1
  [aedffcd0] Static v0.8.7
  [0d7ed370] StaticArrayInterface v1.4.0
  [90137ffa] StaticArrays v1.6.0
  [1e83bf80] StaticArraysCore v1.4.0
  [82ae8749] StatsAPI v1.6.0
  [2913bbd2] StatsBase v0.34.0
  [4c63d2b9] StatsFuns v1.3.0
  [789caeaf] StochasticDiffEq v6.61.1
  [7792a7ef] StrideArraysCore v0.4.16
  [09ab397b] StructArrays v0.6.15
  [53d494c1] StructIO v0.3.0
  [2efcf032] SymbolicIndexingInterface v0.2.2
  [3783bdb8] TableTraits v1.0.1
  [bd369af6] Tables v1.10.1
⌅ [6aa5eb33] TaylorSeries v0.14.0
  [62fd8b95] TensorCore v0.1.1
  [5d786b92] TerminalLoggers v0.1.7
  [8290d209] ThreadingUtilities v0.5.2
  [ac1d9e8a] ThreadsX v0.1.11
  [731e570b] TiffImages v0.6.4
  [a759f4b9] TimerOutputs v0.5.23
  [c751599d] ToeplitzMatrices v0.8.1
  [9f7883ad] Tracker v0.2.26
  [3bb67fe8] TranscodingStreams v0.9.13
  [28d57a85] Transducers v0.4.77
  [592b5752] Trapz v2.0.3
  [a2a6695c] TreeViews v0.3.0
  [d5829a12] TriangularSolve v0.1.19
  [410a4b4d] Tricks v0.1.7
  [981d1d27] TriplotBase v0.1.0
  [781d530d] TruncatedStacktraces v1.3.0
  [9d95972d] TupleTools v1.3.0
  [5c2747f8] URIs v1.4.2
  [3a884ed6] UnPack v1.0.2
  [1cfade01] UnicodeFun v0.4.1
  [1986cc42] Unitful v1.15.0
  [45397f5d] UnitfulLatexify v1.6.3
  [013be700] UnsafeAtomics v0.2.1
⌃ [d80eeb9a] UnsafeAtomicsLLVM v0.1.2
  [41fe7b60] Unzip v0.2.0
  [3d5dd08c] VectorizationBase v0.21.64
  [19fa3120] VertexSafeGraphs v0.2.0
  [efce3f68] WoodburyMatrices v0.5.5
  [e88e6eb3] Zygote v0.6.62
  [700de1a5] ZygoteRules v0.2.3
⌃ [02a925ec] cuDNN v1.0.3
  [6e34b625] Bzip2_jll v1.0.8+0
  [4e9b3aee] CRlibm_jll v1.0.1+0
  [4ee394cb] CUDA_Driver_jll v0.5.0+1
  [76a88914] CUDA_Runtime_jll v0.6.0+0
⌅ [62b44479] CUDNN_jll v8.8.1+0
  [83423d85] Cairo_jll v1.16.1+1
  [7bc98958] Cubature_jll v1.0.5+0
  [5ae413db] EarCut_jll v2.2.4+0
⌅ [7cc45869] Enzyme_jll v0.0.53+0
  [2e619515] Expat_jll v2.5.0+0
  [b22a6f82] FFMPEG_jll v4.4.2+2
  [f5851436] FFTW_jll v3.3.10+0
  [a3f928ae] Fontconfig_jll v2.13.93+0
⌅ [d7e528f0] FreeType2_jll v2.10.4+0
  [559328eb] FriBidi_jll v1.0.10+0
  [0656b61e] GLFW_jll v3.3.8+0
  [d2c73de3] GR_jll v0.72.7+0
  [1b77fbbe] GSL_jll v2.7.2+0
  [78b55507] Gettext_jll v0.21.0+0
  [7746bdde] Glib_jll v2.74.0+2
  [3b182d85] Graphite2_jll v1.3.14+0
  [2e76f6c2] HarfBuzz_jll v2.8.1+1
  [905a6f67] Imath_jll v3.1.7+0
  [1d5cc7b8] IntelOpenMP_jll v2023.1.0+0
  [aacddb02] JpegTurbo_jll v2.1.91+0
  [c1c5ebd0] LAME_jll v3.100.1+0
  [88015f11] LERC_jll v3.0.0+1
⌅ [dad2f222] LLVMExtra_jll v0.0.22+0
  [1d63c593] LLVMOpenMP_jll v15.0.4+0
  [dd4b983a] LZO_jll v2.10.1+0
⌅ [e9f186c6] Libffi_jll v3.2.2+1
  [d4300ac3] Libgcrypt_jll v1.8.7+0
  [7e76a0d4] Libglvnd_jll v1.6.0+0
  [7add5ba3] Libgpg_error_jll v1.42.0+0
  [94ce4f54] Libiconv_jll v1.16.1+2
  [4b2f31a3] Libmount_jll v2.35.0+0
⌅ [89763e89] Libtiff_jll v4.4.0+0
  [38a345b3] Libuuid_jll v2.36.0+0
  [856f044c] MKL_jll v2023.1.0+0
  [e7412a2a] Ogg_jll v1.3.5+1
  [18a262bb] OpenEXR_jll v3.1.4+0
⌅ [458c3c95] OpenSSL_jll v1.1.21+0
  [efe28fd5] OpenSpecFun_jll v0.5.5+0
  [91d4177d] Opus_jll v1.3.2+0
  [36c8627f] Pango_jll v1.50.9+0
  [30392449] Pixman_jll v0.42.2+0
  [460c41e3] QhullMiniWrapper_jll v1.0.0+1
⌅ [784f63db] Qhull_jll v8.0.1003+0
  [ea2cea3b] Qt5Base_jll v5.15.3+2
  [f50d1b31] Rmath_jll v0.4.0+0
  [a2964d1f] Wayland_jll v1.21.0+0
  [2381bf8a] Wayland_protocols_jll v1.25.0+0
  [02c8fc9c] XML2_jll v2.10.3+0
  [aed1982a] XSLT_jll v1.1.34+0
  [4f6342f7] Xorg_libX11_jll v1.8.6+0
  [0c0b7dd1] Xorg_libXau_jll v1.0.11+0
  [935fb764] Xorg_libXcursor_jll v1.2.0+4
  [a3789734] Xorg_libXdmcp_jll v1.1.4+0
  [1082639a] Xorg_libXext_jll v1.3.4+4
  [d091e8ba] Xorg_libXfixes_jll v5.0.3+4
  [a51aa0fd] Xorg_libXi_jll v1.7.10+4
  [d1454406] Xorg_libXinerama_jll v1.1.4+4
  [ec84b674] Xorg_libXrandr_jll v1.5.2+4
  [ea2f1a96] Xorg_libXrender_jll v0.9.10+4
  [14d82f49] Xorg_libpthread_stubs_jll v0.1.1+0
  [c7cfdc94] Xorg_libxcb_jll v1.15.0+0
  [cc61e674] Xorg_libxkbfile_jll v1.1.2+0
  [12413925] Xorg_xcb_util_image_jll v0.4.0+1
  [2def613f] Xorg_xcb_util_jll v0.4.0+1
  [975044d2] Xorg_xcb_util_keysyms_jll v0.4.0+1
  [0d47668e] Xorg_xcb_util_renderutil_jll v0.3.9+1
  [c22f9ab0] Xorg_xcb_util_wm_jll v0.4.1+1
  [35661453] Xorg_xkbcomp_jll v1.4.6+0
  [33bec58e] Xorg_xkeyboard_config_jll v2.39.0+0
  [c5fb5394] Xorg_xtrans_jll v1.5.0+0
  [3161d3a3] Zstd_jll v1.5.5+0
⌅ [214eeab7] fzf_jll v0.29.0+0
  [9a68df92] isoband_jll v0.2.3+0
  [a4ae2306] libaom_jll v3.4.0+0
  [0ac62f75] libass_jll v0.15.1+0
  [f638f0a6] libfdk_aac_jll v2.0.2+0
  [b53b4c65] libpng_jll v1.6.38+0
  [075b6546] libsixel_jll v1.10.3+0
  [f27f6e37] libvorbis_jll v1.3.7+1
  [1270edf5] x264_jll v2021.5.5+0
  [dfaa095f] x265_jll v3.5.0+0
  [d8fb68d0] xkbcommon_jll v1.4.1+0
  [0dad84c5] ArgTools v1.1.1
  [56f22d72] Artifacts
  [2a0f44e3] Base64
  [8bf52ea8] CRC32c
  [ade2ca70] Dates
  [8ba89e20] Distributed
  [f43a241f] Downloads v1.6.0
  [7b1f6079] FileWatching
  [9fa8497b] Future
  [b77e0a4c] InteractiveUtils
  [4af54fe1] LazyArtifacts
  [b27032c2] LibCURL v0.6.3
  [76f85450] LibGit2
  [8f399da3] Libdl
  [37e2e46d] LinearAlgebra
  [56ddb016] Logging
  [d6f4376e] Markdown
  [a63ad114] Mmap
  [ca575930] NetworkOptions v1.2.0
  [44cfe95a] Pkg v1.9.2
  [de0858da] Printf
  [3fa0cd96] REPL
  [9a3f8284] Random
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization
  [1a1011a3] SharedArrays
  [6462fe0b] Sockets
  [2f01184e] SparseArrays
  [10745b16] Statistics v1.9.0
  [4607b0f0] SuiteSparse
  [fa267f1f] TOML v1.0.3
  [a4e569a6] Tar v1.10.0
  [8dfed614] Test
  [cf7118a7] UUIDs
  [4ec0a83e] Unicode
  [e66e0078] CompilerSupportLibraries_jll v1.0.5+0
  [deac9b47] LibCURL_jll v7.84.0+0
  [29816b5a] LibSSH2_jll v1.10.2+0
  [c8ffd9c3] MbedTLS_jll v2.28.2+0
  [14a3606d] MozillaCACerts_jll v2022.10.11
  [4536629a] OpenBLAS_jll v0.3.21+4
  [05823500] OpenLibm_jll v0.8.1+0
  [efcefdf7] PCRE2_jll v10.42.0+0
  [bea87d4a] SuiteSparse_jll v5.10.1+6
  [83775a58] Zlib_jll v1.2.13+0
  [8e850b90] libblastrampoline_jll v5.8.0+0
  [8e850ede] nghttp2_jll v1.48.0+0
  [3f19e933] p7zip_jll v17.4.0+0
Info Packages marked with ⌃ and ⌅ have new versions available, but those with ⌅ are restricted by compatibility constraints from upgrading. To see why use `status --outdated -m`

You can also download the manifest file and the project file.