GlobalSensitivity.jl: Robust, Fast, and Parallel Global Sensitivity Analysis (GSA) in Julia

Global Sensitivity Analysis (GSA) methods are used to quantify the uncertainty in the output of a model with respect to the parameters. These methods allow practitioners to measure both parameter's individual contributions and the contribution of their interactions to the output uncertainty.

Installation

To use this functionality, you must install GlobalSensitivity.jl:

using Pkg
Pkg.add("GlobalSensitivity")

Note: GlobalSensitivity.jl is unrelated to the GlobalSensitivityAnalysis.jl package.

General Interface

The general interface for performing global sensitivity analysis using this package is:

GlobalSensitivity.gsa — Method

gsa(f, method::GSAMethod, param_range; samples, batch=false)

where:

y=f(x) is a function that takes in a single vector and spits out a single vector or scalar. If batch=true, then f takes in a matrix where each row is a set of parameters, and returns a matrix where each row is the output for the corresponding row of parameters.

method is one of the available GSA methods.

param_range is a vector of tuples for the upper and lower bound for the given parameter i.

samples is a required keyword argument for the number of samples of parameters for the design matrix. Note that this is not relevant for Fractional Factorial Method and Morris Method.

Additionally,

For Delta Moment-Independent Method, EASI Method and Regression Method input and output matrix-based method as follows is available:

res = gsa(X, Y, method)

where:

X is the number of parameters * samples matrix with parameter values.

Y is the output dimension * number of samples matrix with are evaluated at X's columns.

method is one of the GSA methods below.

For Sobol Method, one can use the following design matrices-based method instead of parameter range-based method discussed earlier:

effects = gsa(f, method, A, B; batch=false)

where A and B are design matrices, with each row being a set of parameters. Note that generate_design_matrices from QuasiMonteCarlo.jl can be used to generate the design matrices.

The descriptions of the available methods can be found in the Methods section. The gsa interface allows for utilizing batched functions with the batch kwarg discussed above for parallel computation of GSA results.

Citing

If you use this software in your work, please cite:

@article{dixit2022globalsensitivity,
  title={GlobalSensitivity. jl: Performant and Parallel Global Sensitivity Analysis with Julia},
  author={Dixit, Vaibhav Kumar and Rackauckas, Christopher},
  journal={Journal of Open Source Software},
  volume={7},
  number={76},
  pages={4561},
  year={2022}
}

Reproducibility

The documentation of this SciML package was built using these direct dependencies,

Status `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl/docs/Project.toml`
  [13f3f980] CairoMakie v0.15.7
  [ae264745] Copulas v0.1.33
  [31c24e10] Distributions v0.25.122
  [e30172f5] Documenter v1.16.1
  [587475ba] Flux v0.16.5
  [af5da776] GlobalSensitivity v2.8.0 `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl`
  [7f7a1694] Optimization v5.1.0
  [42dfb2eb] OptimizationOptimisers v0.3.15
  [1dea7af3] OrdinaryDiffEq v6.103.0
  [91a5bcdd] Plots v1.41.1
  [8a4e6c94] QuasiMonteCarlo v0.3.3
  [1ed8b502] SciMLSensitivity v7.90.0
  [860ef19b] StableRNGs v1.0.4
  [10745b16] Statistics v1.11.1

and using this machine and Julia version.

Julia Version 1.12.2
Commit ca9b6662be4 (2025-11-20 16:25 UTC)
Build Info:
  Official https://julialang.org release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 4 × AMD EPYC 7763 64-Core Processor
  WORD_SIZE: 64
  LLVM: libLLVM-18.1.7 (ORCJIT, znver3)
  GC: Built with stock GC
Threads: 1 default, 1 interactive, 1 GC (on 4 virtual cores)
Environment:
  JULIA_DEBUG = Documenter

A more complete overview of all dependencies and their versions is also provided.

Status `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl/docs/Manifest.toml`
  [47edcb42] ADTypes v1.19.0
  [a4c015fc] ANSIColoredPrinters v0.0.1
  [621f4979] AbstractFFTs v1.5.0
  [1520ce14] AbstractTrees v0.4.5
  [7d9f7c33] Accessors v0.1.42
  [79e6a3ab] Adapt v4.4.0
  [35492f91] AdaptivePredicates v1.2.0
  [66dad0bd] AliasTables v1.1.3
  [27a7e980] Animations v0.4.2
  [dce04be8] ArgCheck v2.5.0
  [4fba245c] ArrayInterface v7.22.0
  [4c555306] ArrayLayouts v1.12.1
  [a9b6321e] Atomix v1.1.2
  [67c07d97] Automa v1.1.0
  [13072b0f] AxisAlgorithms v1.1.0
  [39de3d68] AxisArrays v0.4.8
  [198e06fe] BangBang v0.4.6
  [18cc8868] BaseDirs v1.3.2
  [9718e550] Baselet v0.1.1
  [0e736298] Bessels v0.2.8
  [7b33fef7] BigCombinatorics v0.3.6
  [d1d4a3ce] BitFlags v0.1.9
  [62783981] BitTwiddlingConvenienceFunctions v0.1.6
  [70df07ce] BracketingNonlinearSolve v1.6.0
  [fa961155] CEnum v0.5.0
  [2a0fbf3d] CPUSummary v0.2.7
  [96374032] CRlibm v1.0.2
  [159f3aea] Cairo v1.1.1
  [13f3f980] CairoMakie v0.15.7
  [7057c7e9] Cassette v0.3.14
  [082447d4] ChainRules v1.72.6
  [d360d2e6] ChainRulesCore v1.26.0
  [fb6a15b2] CloseOpenIntervals v0.1.13
  [944b1d66] CodecZlib v0.7.8
  [a2cac450] ColorBrewer v0.4.2
  [35d6a980] ColorSchemes v3.31.0
  [3da002f7] ColorTypes v0.12.1
  [c3611d14] ColorVectorSpace v0.11.0
  [5ae59095] Colors v0.13.1
  [861a8166] Combinatorics v1.0.3
  [38540f10] CommonSolve v0.2.4
  [bbf7d656] CommonSubexpressions v0.3.1
  [f70d9fcc] CommonWorldInvalidations v1.0.0
  [34da2185] Compat v4.18.1
  [ab4b797d] ComplexityMeasures v3.8.5
  [a33af91c] CompositionsBase v0.1.2
  [95dc2771] ComputePipeline v0.1.5
  [2569d6c7] ConcreteStructs v0.2.3
  [f0e56b4a] ConcurrentUtilities v2.5.0
  [88cd18e8] ConsoleProgressMonitor v0.1.2
  [187b0558] ConstructionBase v1.6.0
  [6add18c4] ContextVariablesX v0.1.3
  [d38c429a] Contour v0.6.3
  [ae264745] Copulas v0.1.33
  [adafc99b] CpuId v0.3.1
  [717857b8] DSP v0.8.4
  [9a962f9c] DataAPI v1.16.0
  [864edb3b] DataStructures v0.19.3
  [e2d170a0] DataValueInterfaces v1.0.0
  [244e2a9f] DefineSingletons v0.1.2
  [927a84f5] DelaunayTriangulation v1.6.5
  [5732040d] DelayEmbeddings v2.9.0
  [8bb1440f] DelimitedFiles v1.9.1
  [2b5f629d] DiffEqBase v6.192.0
  [459566f4] DiffEqCallbacks v4.10.1
  [77a26b50] DiffEqNoiseProcess v5.24.1
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.15.1
  [a0c0ee7d] DifferentiationInterface v0.7.12
  [b4f34e82] Distances v0.10.12
  [31c24e10] Distributions v0.25.122
  [ffbed154] DocStringExtensions v0.9.5
  [e30172f5] Documenter v1.16.1
  [4e289a0a] EnumX v1.0.5
  [7da242da] Enzyme v0.13.107
  [f151be2c] EnzymeCore v0.8.17
  [429591f6] ExactPredicates v2.2.9
  [460bff9d] ExceptionUnwrapping v0.1.11
  [d4d017d3] ExponentialUtilities v1.28.0
  [e2ba6199] ExprTools v0.1.10
  [55351af7] ExproniconLite v0.10.14
  [411431e0] Extents v0.1.6
  [c87230d0] FFMPEG v0.4.5
  [7a1cc6ca] FFTW v1.10.0
  [cc61a311] FLoops v0.2.2
  [b9860ae5] FLoopsBase v0.1.1
  [7034ab61] FastBroadcast v0.3.5
  [9aa1b823] FastClosures v0.3.2
  [442a2c76] FastGaussQuadrature v1.1.0
  [a4df4552] FastPower v1.2.0
  [5789e2e9] FileIO v1.17.1
  [8fc22ac5] FilePaths v0.9.0
  [48062228] FilePathsBase v0.9.24
  [1a297f60] FillArrays v1.15.0
  [6a86dc24] FiniteDiff v2.29.0
  [53c48c17] FixedPointNumbers v0.8.5
  [587475ba] Flux v0.16.5
  [1fa38f19] Format v1.3.7
  [f6369f11] ForwardDiff v1.3.0
  [b38be410] FreeType v4.1.1
  [663a7486] FreeTypeAbstraction v0.10.8
  [f62d2435] FunctionProperties v0.1.2
  [069b7b12] FunctionWrappers v1.1.3
  [77dc65aa] FunctionWrappersWrappers v0.1.3
  [d9f16b24] Functors v0.5.2
  [46192b85] GPUArraysCore v0.2.0
  [61eb1bfa] GPUCompiler v1.7.4
  [28b8d3ca] GR v0.73.19
  [c145ed77] GenericSchur v0.5.6
  [5c1252a2] GeometryBasics v0.5.10
  [d7ba0133] Git v1.5.0
  [af5da776] GlobalSensitivity v2.8.0 `~/work/GlobalSensitivity.jl/GlobalSensitivity.jl`
  [a2bd30eb] Graphics v1.1.3
  [3955a311] GridLayoutBase v0.11.2
  [42e2da0e] Grisu v1.0.2
  [19dc6840] HCubature v1.8.0
  [cd3eb016] HTTP v1.10.19
  [076d061b] HashArrayMappedTries v0.2.0
  [34004b35] HypergeometricFunctions v0.3.28
  [b5f81e59] IOCapture v1.0.0
  [7869d1d1] IRTools v0.4.15
  [615f187c] IfElse v0.1.1
  [2803e5a7] ImageAxes v0.6.12
  [c817782e] ImageBase v0.1.7
  [a09fc81d] ImageCore v0.10.5
  [82e4d734] ImageIO v0.6.9
  [bc367c6b] ImageMetadata v0.9.10
  [9b13fd28] IndirectArrays v1.0.0
  [d25df0c9] Inflate v0.1.5
  [22cec73e] InitialValues v0.3.1
  [18e54dd8] IntegerMathUtils v0.1.3
  [a98d9a8b] Interpolations v0.16.2
  [d1acc4aa] IntervalArithmetic v1.0.2
  [8197267c] IntervalSets v0.7.13
  [3587e190] InverseFunctions v0.1.17
  [92d709cd] IrrationalConstants v0.2.6
  [f1662d9f] Isoband v0.1.1
  [c8e1da08] IterTools v1.10.0
  [82899510] IteratorInterfaceExtensions v1.0.0
  [1019f520] JLFzf v0.1.11
  [692b3bcd] JLLWrappers v1.7.1
  [682c06a0] JSON v1.3.0
  [ae98c720] Jieko v0.2.1
  [b835a17e] JpegTurbo v0.1.6
  [b14d175d] JuliaVariables v0.2.4
  [63c18a36] KernelAbstractions v0.9.39
  [5ab0869b] KernelDensity v0.6.10
  [ba0b0d4f] Krylov v0.10.3
  [929cbde3] LLVM v9.4.4
  [b964fa9f] LaTeXStrings v1.4.0
  [984bce1d] LambertW v1.0.0
  [23fbe1c1] Latexify v0.16.10
  [73f95e8e] LatticeRules v0.0.1
  [10f19ff3] LayoutPointers v0.1.17
  [0e77f7df] LazilyInitializedFields v1.3.0
  [5078a376] LazyArrays v2.9.1
  [8cdb02fc] LazyModules v0.3.1
  [1d6d02ad] LeftChildRightSiblingTrees v0.2.1
  [87fe0de2] LineSearch v0.1.4
  [d3d80556] LineSearches v7.4.1
  [7ed4a6bd] LinearSolve v3.47.0
  [2ab3a3ac] LogExpFunctions v0.3.29
  [e6f89c97] LoggingExtras v1.2.0
  [c2834f40] MLCore v1.0.0
  [7e8f7934] MLDataDevices v1.15.1
  [d8e11817] MLStyle v0.4.17
  [f1d291b0] MLUtils v0.4.8
  [1914dd2f] MacroTools v0.5.16
  [ee78f7c6] Makie v0.24.7
  [d125e4d3] ManualMemory v0.1.8
  [dbb5928d] MappedArrays v0.4.2
  [d0879d2d] MarkdownAST v0.1.2
  [0a4f8689] MathTeXEngine v0.6.7
  [bb5d69b7] MaybeInplace v0.1.4
  [739be429] MbedTLS v1.1.9
  [442fdcdd] Measures v0.3.3
  [128add7d] MicroCollections v0.2.0
  [e1d29d7a] Missings v1.2.0
  [e94cdb99] MosaicViews v0.3.4
  [2e0e35c7] Moshi v0.3.7
  [46d2c3a1] MuladdMacro v0.2.4
  [37188c8d] MvNormalCDF v0.3.2
  [d41bc354] NLSolversBase v7.10.0
  [872c559c] NNlib v0.9.31
  [77ba4419] NaNMath v1.1.3
  [71a1bf82] NameResolution v0.1.5
  [b8a86587] NearestNeighbors v0.4.24
  [645ca80c] Neighborhood v0.2.4
  [f09324ee] Netpbm v1.1.1
  [8913a72c] NonlinearSolve v4.12.0
  [be0214bd] NonlinearSolveBase v2.4.0
  [5959db7a] NonlinearSolveFirstOrder v1.10.0
  [9a2c21bd] NonlinearSolveQuasiNewton v1.11.0
  [26075421] NonlinearSolveSpectralMethods v1.6.0
  [d8793406] ObjectFile v0.5.0
  [510215fc] Observables v0.5.5
  [6fe1bfb0] OffsetArrays v1.17.0
  [0b1bfda6] OneHotArrays v0.2.10
  [52e1d378] OpenEXR v0.3.3
  [4d8831e6] OpenSSL v1.6.0
  [429524aa] Optim v1.13.2
  [3bd65402] Optimisers v0.4.6
  [7f7a1694] Optimization v5.1.0
  [bca83a33] OptimizationBase v4.0.2
  [42dfb2eb] OptimizationOptimisers v0.3.15
  [bac558e1] OrderedCollections v1.8.1
  [1dea7af3] OrdinaryDiffEq v6.103.0
  [89bda076] OrdinaryDiffEqAdamsBashforthMoulton v1.5.0
  [6ad6398a] OrdinaryDiffEqBDF v1.10.1
  [bbf590c4] OrdinaryDiffEqCore v1.36.0
  [50262376] OrdinaryDiffEqDefault v1.8.0
  [4302a76b] OrdinaryDiffEqDifferentiation v1.16.1
  [9286f039] OrdinaryDiffEqExplicitRK v1.4.0
  [e0540318] OrdinaryDiffEqExponentialRK v1.8.0
  [becaefa8] OrdinaryDiffEqExtrapolation v1.9.0
  [5960d6e9] OrdinaryDiffEqFIRK v1.16.0
  [101fe9f7] OrdinaryDiffEqFeagin v1.4.0
  [d3585ca7] OrdinaryDiffEqFunctionMap v1.5.0
  [d28bc4f8] OrdinaryDiffEqHighOrderRK v1.5.0
  [9f002381] OrdinaryDiffEqIMEXMultistep v1.7.0
  [521117fe] OrdinaryDiffEqLinear v1.6.0
  [1344f307] OrdinaryDiffEqLowOrderRK v1.6.0
  [b0944070] OrdinaryDiffEqLowStorageRK v1.7.0
  [127b3ac7] OrdinaryDiffEqNonlinearSolve v1.15.0
  [c9986a66] OrdinaryDiffEqNordsieck v1.4.0
  [5dd0a6cf] OrdinaryDiffEqPDIRK v1.6.0
  [5b33eab2] OrdinaryDiffEqPRK v1.4.0
  [04162be5] OrdinaryDiffEqQPRK v1.4.0
  [af6ede74] OrdinaryDiffEqRKN v1.5.0
  [43230ef6] OrdinaryDiffEqRosenbrock v1.18.1
  [2d112036] OrdinaryDiffEqSDIRK v1.7.0
  [669c94d9] OrdinaryDiffEqSSPRK v1.7.0
  [e3e12d00] OrdinaryDiffEqStabilizedIRK v1.6.0
  [358294b1] OrdinaryDiffEqStabilizedRK v1.4.0
  [fa646aed] OrdinaryDiffEqSymplecticRK v1.7.0
  [b1df2697] OrdinaryDiffEqTsit5 v1.5.0
  [79d7bb75] OrdinaryDiffEqVerner v1.6.0
  [90014a1f] PDMats v0.11.36
  [f57f5aa1] PNGFiles v0.4.4
  [19eb6ba3] Packing v0.5.1
  [5432bcbf] PaddedViews v0.5.12
  [d96e819e] Parameters v0.12.3
  [69de0a69] Parsers v2.8.3
  [eebad327] PkgVersion v0.3.3
  [ccf2f8ad] PlotThemes v3.3.0
  [995b91a9] PlotUtils v1.4.4
  [91a5bcdd] Plots v1.41.1
  [e409e4f3] PoissonRandom v0.4.7
  [85e3b03c] PolyLog v2.6.0
  [f517fe37] Polyester v0.7.18
  [1d0040c9] PolyesterWeave v0.2.2
  [647866c9] PolygonOps v0.1.2
  [f27b6e38] Polynomials v4.1.0
  [85a6dd25] PositiveFactorizations v0.2.4
  [d236fae5] PreallocationTools v0.4.34
  [aea7be01] PrecompileTools v1.3.3
  [21216c6a] Preferences v1.5.0
  [8162dcfd] PrettyPrint v0.2.0
  [27ebfcd6] Primes v0.5.7
  [33c8b6b6] ProgressLogging v0.1.6
  [92933f4c] ProgressMeter v1.11.0
  [43287f4e] PtrArrays v1.3.0
  [4b34888f] QOI v1.0.1
  [1fd47b50] QuadGK v2.11.2
  [8a4e6c94] QuasiMonteCarlo v0.3.3
  [74087812] Random123 v1.7.1
  [e6cf234a] RandomNumbers v1.6.0
  [b3c3ace0] RangeArrays v0.3.2
  [c84ed2f1] Ratios v0.4.5
  [c1ae055f] RealDot v0.1.0
  [3cdcf5f2] RecipesBase v1.3.4
  [01d81517] RecipesPipeline v0.6.12
  [731186ca] RecursiveArrayTools v3.39.0
  [189a3867] Reexport v1.2.2
  [42d2dcc6] Referenceables v0.1.3
  [2792f1a3] RegistryInstances v0.1.0
  [05181044] RelocatableFolders v1.0.1
  [ae029012] Requires v1.3.1
  [ae5879a3] ResettableStacks v1.1.1
  [37e2e3b7] ReverseDiff v1.16.1
  [79098fc4] Rmath v0.9.0
  [f2b01f46] Roots v2.2.10
  [5eaf0fd0] RoundingEmulator v0.2.1
  [7e49a35a] RuntimeGeneratedFunctions v0.5.16
  [fdea26ae] SIMD v3.7.2
  [94e857df] SIMDTypes v0.1.0
  [0bca4576] SciMLBase v2.127.0
  [19f34311] SciMLJacobianOperators v0.1.11
  [a6db7da4] SciMLLogging v1.5.0
  [c0aeaf25] SciMLOperators v1.13.0
  [431bcebd] SciMLPublic v1.0.0
  [1ed8b502] SciMLSensitivity v7.90.0
  [53ae85a6] SciMLStructures v1.7.0
  [7e506255] ScopedValues v1.5.0
  [6c6a2e73] Scratch v1.3.0
  [efcf1570] Setfield v1.1.2
  [65257c39] ShaderAbstractions v0.5.0
  [605ecd9f] ShowCases v0.1.0
  [992d4aef] Showoff v1.0.3
  [73760f76] SignedDistanceFields v0.4.0
  [777ac1f9] SimpleBufferStream v1.2.0
  [727e6d20] SimpleNonlinearSolve v2.9.0
  [699a6c99] SimpleTraits v0.9.5
  [ce78b400] SimpleUnPack v1.1.0
  [45858cf5] Sixel v0.1.5
  [ed01d8cd] Sobol v1.5.0
  [a2af1166] SortingAlgorithms v1.2.2
  [9f842d2f] SparseConnectivityTracer v1.1.3
  [dc90abb0] SparseInverseSubset v0.1.2
  [0a514795] SparseMatrixColorings v0.4.23
  [276daf66] SpecialFunctions v2.6.1
  [171d559e] SplittablesBase v0.1.15
  [860ef19b] StableRNGs v1.0.4
  [cae243ae] StackViews v0.1.2
  [40b095a5] StateSpaceSets v2.5.3
  [aedffcd0] Static v1.3.1
  [0d7ed370] StaticArrayInterface v1.8.0
  [90137ffa] StaticArrays v1.9.15
  [1e83bf80] StaticArraysCore v1.4.4
  [10745b16] Statistics v1.11.1
  [82ae8749] StatsAPI v1.7.1
  [2913bbd2] StatsBase v0.34.8
  [4c63d2b9] StatsFuns v1.5.2
  [7792a7ef] StrideArraysCore v0.5.8
  [09ab397b] StructArrays v0.7.2
  [53d494c1] StructIO v0.3.1
  [ec057cc2] StructUtils v2.6.0
  [2efcf032] SymbolicIndexingInterface v0.3.46
  [3783bdb8] TableTraits v1.0.1
  [bd369af6] Tables v1.12.1
  [6aa5eb33] TaylorSeries v0.20.9
  [62fd8b95] TensorCore v0.1.1
  [5d786b92] TerminalLoggers v0.1.7
  [8290d209] ThreadingUtilities v0.5.5
  [ac1d9e8a] ThreadsX v0.1.12
  [731e570b] TiffImages v0.11.6
  [a759f4b9] TimerOutputs v0.5.29
  [9f7883ad] Tracker v0.2.38
  [e689c965] Tracy v0.1.6
  [3bb67fe8] TranscodingStreams v0.11.3
  [28d57a85] Transducers v0.4.85
  [592b5752] Trapz v2.0.3
  [981d1d27] TriplotBase v0.1.0
  [781d530d] TruncatedStacktraces v1.4.0
  [5c2747f8] URIs v1.6.1
  [3a884ed6] UnPack v1.0.2
  [1cfade01] UnicodeFun v0.4.1
  [1986cc42] Unitful v1.25.1
  [013be700] UnsafeAtomics v0.3.0
  [41fe7b60] Unzip v0.2.0
  [29a6e085] Wavelets v0.10.1
  [e3aaa7dc] WebP v0.1.3
  [efce3f68] WoodburyMatrices v1.0.0
  [e88e6eb3] Zygote v0.7.10
  [700de1a5] ZygoteRules v0.2.7
  [6e34b625] Bzip2_jll v1.0.9+0
  [4e9b3aee] CRlibm_jll v1.0.1+0
  [83423d85] Cairo_jll v1.18.5+0
  [ee1fde0b] Dbus_jll v1.16.2+0
  [5ae413db] EarCut_jll v2.2.4+0
  [7cc45869] Enzyme_jll v0.0.220+0
  [2702e6a9] EpollShim_jll v0.0.20230411+1
  [2e619515] Expat_jll v2.7.3+0
⌅ [b22a6f82] FFMPEG_jll v7.1.1+0
  [f5851436] FFTW_jll v3.3.11+0
  [a3f928ae] Fontconfig_jll v2.17.1+0
  [d7e528f0] FreeType2_jll v2.13.4+0
  [559328eb] FriBidi_jll v1.0.17+0
  [0656b61e] GLFW_jll v3.4.0+2
  [d2c73de3] GR_jll v0.73.19+1
  [b0724c58] GettextRuntime_jll v0.22.4+0
  [61579ee1] Ghostscript_jll v9.55.1+0
  [59f7168a] Giflib_jll v5.2.3+0
  [020c3dae] Git_LFS_jll v3.7.0+0
  [f8c6e375] Git_jll v2.51.3+0
  [7746bdde] Glib_jll v2.86.0+0
  [3b182d85] Graphite2_jll v1.3.15+0
  [2e76f6c2] HarfBuzz_jll v8.5.1+0
  [905a6f67] Imath_jll v3.1.11+0
  [1d5cc7b8] IntelOpenMP_jll v2025.2.0+0
  [aacddb02] JpegTurbo_jll v3.1.3+0
  [c1c5ebd0] LAME_jll v3.100.3+0
  [88015f11] LERC_jll v4.0.1+0
  [dad2f222] LLVMExtra_jll v0.0.38+0
  [1d63c593] LLVMOpenMP_jll v18.1.8+0
  [dd4b983a] LZO_jll v2.10.3+0
  [ad6e5548] LibTracyClient_jll v0.9.1+6
⌅ [e9f186c6] Libffi_jll v3.4.7+0
  [7e76a0d4] Libglvnd_jll v1.7.1+1
  [94ce4f54] Libiconv_jll v1.18.0+0
  [4b2f31a3] Libmount_jll v2.41.2+0
  [89763e89] Libtiff_jll v4.7.2+0
  [38a345b3] Libuuid_jll v2.41.2+0
  [856f044c] MKL_jll v2025.2.0+0
  [c8ffd9c3] MbedTLS_jll v2.28.1010+0
  [e7412a2a] Ogg_jll v1.3.6+0
  [6cdc7f73] OpenBLASConsistentFPCSR_jll v0.3.29+0
  [18a262bb] OpenEXR_jll v3.2.4+0
  [9bd350c2] OpenSSH_jll v10.2.1+0
  [efe28fd5] OpenSpecFun_jll v0.5.6+0
  [91d4177d] Opus_jll v1.5.2+0
  [36c8627f] Pango_jll v1.57.0+0
⌅ [30392449] Pixman_jll v0.44.2+0
  [c0090381] Qt6Base_jll v6.8.2+2
  [629bc702] Qt6Declarative_jll v6.8.2+1
  [ce943373] Qt6ShaderTools_jll v6.8.2+1
  [e99dba38] Qt6Wayland_jll v6.8.2+2
  [f50d1b31] Rmath_jll v0.5.1+0
  [a44049a8] Vulkan_Loader_jll v1.3.243+0
  [a2964d1f] Wayland_jll v1.24.0+0
  [ffd25f8a] XZ_jll v5.8.1+0
  [f67eecfb] Xorg_libICE_jll v1.1.2+0
  [c834827a] Xorg_libSM_jll v1.2.6+0
  [4f6342f7] Xorg_libX11_jll v1.8.12+0
  [0c0b7dd1] Xorg_libXau_jll v1.0.13+0
  [935fb764] Xorg_libXcursor_jll v1.2.4+0
  [a3789734] Xorg_libXdmcp_jll v1.1.6+0
  [1082639a] Xorg_libXext_jll v1.3.7+0
  [d091e8ba] Xorg_libXfixes_jll v6.0.2+0
  [a51aa0fd] Xorg_libXi_jll v1.8.3+0
  [d1454406] Xorg_libXinerama_jll v1.1.6+0
  [ec84b674] Xorg_libXrandr_jll v1.5.5+0
  [ea2f1a96] Xorg_libXrender_jll v0.9.12+0
  [c7cfdc94] Xorg_libxcb_jll v1.17.1+0
  [cc61e674] Xorg_libxkbfile_jll v1.1.3+0
  [e920d4aa] Xorg_xcb_util_cursor_jll v0.1.6+0
  [12413925] Xorg_xcb_util_image_jll v0.4.1+0
  [2def613f] Xorg_xcb_util_jll v0.4.1+0
  [975044d2] Xorg_xcb_util_keysyms_jll v0.4.1+0
  [0d47668e] Xorg_xcb_util_renderutil_jll v0.3.10+0
  [c22f9ab0] Xorg_xcb_util_wm_jll v0.4.2+0
  [35661453] Xorg_xkbcomp_jll v1.4.7+0
  [33bec58e] Xorg_xkeyboard_config_jll v2.44.0+0
  [c5fb5394] Xorg_xtrans_jll v1.6.0+0
  [3161d3a3] Zstd_jll v1.5.7+1
  [35ca27e7] eudev_jll v3.2.14+0
  [214eeab7] fzf_jll v0.61.1+0
  [9a68df92] isoband_jll v0.2.3+0
  [a4ae2306] libaom_jll v3.13.1+0
  [0ac62f75] libass_jll v0.17.4+0
  [1183f4f0] libdecor_jll v0.2.2+0
  [2db6ffa8] libevdev_jll v1.13.4+0
  [f638f0a6] libfdk_aac_jll v2.0.4+0
  [36db933b] libinput_jll v1.28.1+0
  [b53b4c65] libpng_jll v1.6.50+0
  [075b6546] libsixel_jll v1.10.5+0
  [f27f6e37] libvorbis_jll v1.3.8+0
  [c5f90fcd] libwebp_jll v1.6.0+0
  [009596ad] mtdev_jll v1.1.7+0
  [1317d2d5] oneTBB_jll v2022.0.0+1
  [1270edf5] x264_jll v10164.0.1+0
  [dfaa095f] x265_jll v4.1.0+0
  [d8fb68d0] xkbcommon_jll v1.9.2+0
  [0dad84c5] ArgTools v1.1.2
  [56f22d72] Artifacts v1.11.0
  [2a0f44e3] Base64 v1.11.0
  [8bf52ea8] CRC32c v1.11.0
  [ade2ca70] Dates v1.11.0
  [8ba89e20] Distributed v1.11.0
  [f43a241f] Downloads v1.7.0
  [7b1f6079] FileWatching v1.11.0
  [9fa8497b] Future v1.11.0
  [b77e0a4c] InteractiveUtils v1.11.0
  [ac6e5ff7] JuliaSyntaxHighlighting v1.12.0
  [4af54fe1] LazyArtifacts v1.11.0
  [b27032c2] LibCURL v0.6.4
  [76f85450] LibGit2 v1.11.0
  [8f399da3] Libdl v1.11.0
  [37e2e46d] LinearAlgebra v1.12.0
  [56ddb016] Logging v1.11.0
  [d6f4376e] Markdown v1.11.0
  [a63ad114] Mmap v1.11.0
  [ca575930] NetworkOptions v1.3.0
  [44cfe95a] Pkg v1.12.0
  [de0858da] Printf v1.11.0
  [3fa0cd96] REPL v1.11.0
  [9a3f8284] Random v1.11.0
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization v1.11.0
  [1a1011a3] SharedArrays v1.11.0
  [6462fe0b] Sockets v1.11.0
  [2f01184e] SparseArrays v1.12.0
  [f489334b] StyledStrings v1.11.0
  [4607b0f0] SuiteSparse
  [fa267f1f] TOML v1.0.3
  [a4e569a6] Tar v1.10.0
  [8dfed614] Test v1.11.0
  [cf7118a7] UUIDs v1.11.0
  [4ec0a83e] Unicode v1.11.0
  [e66e0078] CompilerSupportLibraries_jll v1.3.0+1
  [deac9b47] LibCURL_jll v8.15.0+0
  [e37daf67] LibGit2_jll v1.9.0+0
  [29816b5a] LibSSH2_jll v1.11.3+1
  [14a3606d] MozillaCACerts_jll v2025.5.20
  [4536629a] OpenBLAS_jll v0.3.29+0
  [05823500] OpenLibm_jll v0.8.7+0
  [458c3c95] OpenSSL_jll v3.5.4+0
  [efcefdf7] PCRE2_jll v10.44.0+1
  [bea87d4a] SuiteSparse_jll v7.8.3+2
  [83775a58] Zlib_jll v1.3.1+2
  [8e850b90] libblastrampoline_jll v5.15.0+0
  [8e850ede] nghttp2_jll v1.64.0+1
  [3f19e933] p7zip_jll v17.7.0+0
Info Packages marked with ⌅ have new versions available but compatibility constraints restrict them from upgrading. To see why use `status --outdated -m`

You can also download the manifest file and the project file.